Share Vite-entry hydration logic via lib/hydrate_entry.jsx
Problem: The three Vite entry files (app.entry.jsx, chrome.entry.jsx,
sidebar.entry.jsx) each repeated ~30 lines of identical
`readBoot` / `mount` / DOMContentLoaded plumbing, differing only in
the registry import. Any future change to the boot contract had to be
made in three places, inviting drift.

Fix: factor the shared hydration logic into
`lib/hydrate_entry.jsx::hydrateRouteFromBoot`. Each entry file shrinks
to a 3-line shim that imports the helper and its per-bundle registry
resolver. The Vite build still emits three separate bundles (one per
rollup input) because the per-bundle imports remain in their own
entries.
